\name{compute_percent_mito}
\alias{compute_percent_mito}
\title{compute_percent_mito}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
compute_percent_mito(so, mitoGenes)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for compute_percent_mito workflow examples.
}
