\name{density_plot_complexity}
\alias{density_plot_complexity}
\title{density_plot_complexity}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
density_plot_complexity(so, plot_nrows)
}
\arguments{
  \item{so}{Argument accepted by \code{density_plot_complexity}.}
  \item{plot_nrows}{Argument accepted by \code{density_plot_complexity}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for density_plot_complexity workflow examples.
}
