\name{density_plot_n_umis}
\alias{density_plot_n_umis}
\title{density_plot_n_umis}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
density_plot_n_umis(so, plot_nrows)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for density_plot_n_umis workflow examples.
}
