\name{plot_highest_expressed}
\alias{plot_highest_expressed}
\title{plot_highest_expressed}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
plot_highest_expressed(so, nfeat2plot, boxplot_alpha, ylab_txt, colors)
}
\arguments{
  \item{so}{Argument accepted by \code{plot_highest_expressed}.}
  \item{nfeat2plot}{Argument accepted by \code{plot_highest_expressed}.}
  \item{boxplot_alpha}{Argument accepted by \code{plot_highest_expressed}.}
  \item{ylab_txt}{Argument accepted by \code{plot_highest_expressed}.}
  \item{colors}{Argument accepted by \code{plot_highest_expressed}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for plot_highest_expressed workflow examples.
}
