\name{prepare_so_merging_Seurat4}
\alias{prepare_so_merging_Seurat4}
\title{prepare_so_merging_Seurat4}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
prepare_so_merging_Seurat4(so, verbose)
}
\arguments{
  \item{so}{Argument accepted by \code{prepare_so_merging_Seurat4}.}
  \item{verbose}{Argument accepted by \code{prepare_so_merging_Seurat4}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for prepare_so_merging_Seurat4 workflow examples.
}
