\name{qc_distri_plot}
\alias{qc_distri_plot}
\title{qc_distri_plot}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
qc_distri_plot(so, measure)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for qc_distri_plot workflow examples.
}
