\name{scater_qc}
\alias{scater_qc}
\title{scater_qc}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
scater_qc(so)
}
\arguments{
  \item{so}{Argument accepted by \code{scater_qc}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for scater_qc workflow examples.
}
