\name{split_print_legend}
\alias{split_print_legend}
\title{split_print_legend}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
split_print_legend(plt)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for split_print_legend workflow examples.
}
