\name{baia.seurat.helpers-package}
\alias{baia.seurat.helpers}
\alias{baia.seurat.helpers-package}
\title{Reusable Seurat helper functions for BAIA scRNA-seq analyses}
\description{
Provides reusable Seurat-centric quality-control, clustering, merging, sample-sheet, and plotting helpers used across BAIA scRNA-seq projects.
}
\docType{package}
