\name{boxplot_n_genes}
\alias{boxplot_n_genes}
\title{boxplot_n_genes}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
boxplot_n_genes(so)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for boxplot_n_genes workflow examples.
}
