\name{boxplot_n_umis}
\alias{boxplot_n_umis}
\title{boxplot_n_umis}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
boxplot_n_umis(so)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for boxplot_n_umis workflow examples.
}
