\name{density_plot_mito_ratio}
\alias{density_plot_mito_ratio}
\title{density_plot_mito_ratio}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
density_plot_mito_ratio(so, plot_nrows)
}
\arguments{
  \item{so}{Argument accepted by \code{density_plot_mito_ratio}.}
  \item{plot_nrows}{Argument accepted by \code{density_plot_mito_ratio}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for density_plot_mito_ratio workflow examples.
}
