\name{density_plot_n_genes}
\alias{density_plot_n_genes}
\title{density_plot_n_genes}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
density_plot_n_genes(so, plot_nrows)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for density_plot_n_genes workflow examples.
}
