\name{dotplot_n_umis_genes_mito}
\alias{dotplot_n_umis_genes_mito}
\title{dotplot_n_umis_genes_mito}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
dotplot_n_umis_genes_mito(so, plot_nrows)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for dotplot_n_umis_genes_mito workflow examples.
}
