\name{factor_resolution}
\alias{factor_resolution}
\title{factor_resolution}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
factor_resolution(so)
}
\arguments{
  \item{so}{Argument accepted by \code{factor_resolution}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for factor_resolution workflow examples.
}
