\name{histogram_n_cells}
\alias{histogram_n_cells}
\title{histogram_n_cells}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
histogram_n_cells(so)
}
\arguments{
  \item{so}{Argument accepted by \code{histogram_n_cells}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for histogram_n_cells workflow examples.
}
