\name{plot_basic_qc}
\alias{plot_basic_qc}
\title{plot_basic_qc}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
plot_basic_qc(so, title, colors, alpha)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for plot_basic_qc workflow examples.
}
