\name{prepare_so_merging}
\alias{prepare_so_merging}
\title{prepare_so_merging}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
prepare_so_merging(so)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for prepare_so_merging workflow examples.
}
