\name{show_qc_plots}
\alias{show_qc_plots}
\title{show_qc_plots}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
show_qc_plots(so, plot_nrows)
}
\arguments{
  \item{so}{Argument accepted by \code{show_qc_plots}.}
  \item{plot_nrows}{Argument accepted by \code{show_qc_plots}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for show_qc_plots workflow examples.
}
