\name{so_version}
\alias{so_version}
\title{so_version}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
so_version(so)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for so_version workflow examples.
}
