\name{test_cell_cycle_effect}
\alias{test_cell_cycle_effect}
\title{test_cell_cycle_effect}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
test_cell_cycle_effect(so, so_label, s_genes, g2m_genes, verbose, showplots)
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for test_cell_cycle_effect workflow examples.
}
