\name{test_cell_cycle_effect}
\alias{test_cell_cycle_effect}
\title{test_cell_cycle_effect}
\description{
Reusable helper exported by Reusable Seurat helper functions for BAIA scRNA-seq analyses.
}
\usage{
test_cell_cycle_effect(so, so_label, s_genes, g2m_genes, verbose, showplots)
}
\arguments{
  \item{so}{Argument accepted by \code{test_cell_cycle_effect}.}
  \item{so_label}{Argument accepted by \code{test_cell_cycle_effect}.}
  \item{s_genes}{Argument accepted by \code{test_cell_cycle_effect}.}
  \item{g2m_genes}{Argument accepted by \code{test_cell_cycle_effect}.}
  \item{verbose}{Argument accepted by \code{test_cell_cycle_effect}.}
  \item{showplots}{Argument accepted by \code{test_cell_cycle_effect}.}
}
\value{
An R object returned by the helper.
}
\examples{
# See the package vignettes for test_cell_cycle_effect workflow examples.
}
